Characteristics and parseable UUID tests added, first integration test is under development

+++ src/main/java/co/edu/icesi/zoo/dto/AnimalDTO.java
@@ -25,7 +25,7 @@ public class AnimalDTO {
     private String name;
     private char sex;
     private double weight;
-    private int age;
+    private double age;
     private double height;
     private String arrivalDate;
     private String motherId;
