+++ src/main/java/co/edu/icesi/zoo/dto/AnimalNoParentsDTO.java
@@ -17,7 +17,7 @@ public class AnimalNoParentsDTO {
     private String name;
     private char sex;
     private double weight;
-    private int age;
+    private double age;
     private double height;
     private String arrivalDate;
